```
 name    old time/op    new time/op    delta
 Dump-8     5.02s ± 2%     3.98s ± 2%  -20.83%  (p=0.000 n=29+29)

 name    old alloc/op   new alloc/op   delta
 Dump-8    2.50GB ± 0%    2.40GB ± 0%   -3.82%  (p=0.000 n=29+30)

 name    old allocs/op  new allocs/op  delta
 Dump-8     38.8M ± 0%     29.3M ± 0%  -24.69%  (p=0.000 n=30+29)
```